+++ movies/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
     </dist:module>
 
     <application>
-        <activity android:name=".ui.views.MovieActivity">
+        <activity android:name=".ui.views.activities.MovieActivity">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
             </intent-filter>
