+++ movies/src/main/java/com/androidstudy/movies/data/datastates/NetworkResult.kt
@@ -1,4 +1,4 @@
-package com.androidstudy.movies.ui.views.datastates
+package com.androidstudy.movies.data.datastates
 
 sealed class NetworkResult<out T : Any> {
     data class Success<out T : Any>(val data: T) : NetworkResult<T>()
