+++ movies/src/main/java/com/androidstudy/movies/data/models/Character.kt
@@ -1,4 +1,4 @@
-package com.androidstudy.movies.ui.views.models
+package com.androidstudy.movies.data.models
 
 data class Character(
     val created: String,
