+++ movies/src/main/java/com/androidstudy/movies/data/models/Origin.kt
@@ -1,4 +1,4 @@
-package com.androidstudy.movies.ui.views.models
+package com.androidstudy.movies.data.models
 
 data class Origin(
     val name: String,
