+++ movies/src/main/java/com/androidstudy/movies/data/network/ApiService.kt
@@ -1,6 +1,6 @@
-package com.androidstudy.movies.ui.views.api
+package com.androidstudy.movies.data.network
 
-import com.androidstudy.movies.ui.views.models.CharactersResponseModel
+import com.androidstudy.movies.data.models.CharactersResponseModel
 import retrofit2.Response
 import retrofit2.http.GET
