+++ movies/src/main/java/com/androidstudy/movies/ui/adapter/CharactersAdapater.kt
@@ -1,4 +1,4 @@
-package com.androidstudy.movies.ui.views.ui.characters
+package com.androidstudy.movies.ui.adapter
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import coil.api.load
 import com.androidstudy.movies.R
-import com.androidstudy.movies.ui.views.models.Character
+import com.androidstudy.movies.data.models.Character
 import kotlinx.android.synthetic.main.row_movie_item.view.*
 
 typealias  ClickListener = (Character) -> Unit
@@ -19,7 +19,10 @@ class CharactersAdapater(
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CharactersViewHolder {
         val itemView = LayoutInflater.from(parent.context)
             .inflate(R.layout.row_movie_item, parent, false)
-        return CharactersViewHolder(itemView, clickListener)
+        return CharactersViewHolder(
+            itemView,
+            clickListener
+        )
     }
 
     override fun getItemCount(): Int = charactersList.size
