--- movies/src/main/java/com/androidstudy/movies/ui/views/api/ApiEndpoints.kt
@@ -1,5 +1,0 @@
-package com.androidstudy.movies.ui.views.api
-
-object ApiEndpoints {
-    const val BASE_URL = "https://rickandmortyapi.com/api/"
-}
