+++ movies/src/main/java/com/androidstudy/movies/ui/views/fragments/characterdetails/CharacterDetailsFragment.kt
@@ -1,0 +1,6 @@
+package com.androidstudy.movies.ui.views.fragments.characterdetails
+
+import androidx.fragment.app.Fragment
+import com.androidstudy.movies.R
+
+class CharacterDetailsFragment : Fragment(R.layout.fragment_character_details)
