--- movies/src/main/java/com/androidstudy/movies/ui/views/ui/characterdetails/CharacterDetailsFragment.kt
@@ -1,7 +1,0 @@
-package com.androidstudy.movies.ui.views.ui.characterdetails
-
-import androidx.fragment.app.Fragment
-import com.androidstudy.movies.R
-
-class CharacterDetailsFragment : Fragment(R.layout.fragment_character_details) {
-}
