--- movies/src/main/java/com/androidstudy/movies/ui/views/utils/NonNullMediatorLiveData.kt
@@ -1,6 +1,0 @@
-package com.androidstudy.movies.ui.views.utils
-
-import androidx.lifecycle.MediatorLiveData
-
-class NonNullMediatorLiveData<T> : MediatorLiveData<T>() {
-}
