+++ movies/src/main/java/com/androidstudy/movies/utils/NonNullMediatorLiveData.kt
@@ -1,0 +1,5 @@
+package com.androidstudy.movies.utils
+
+import androidx.lifecycle.MediatorLiveData
+
+class NonNullMediatorLiveData<T> : MediatorLiveData<T>()
