+++ movies/src/main/java/com/androidstudy/movies/utils/Utils.kt
@@ -1,0 +1,5 @@
+package com.androidstudy.movies.utils
+
+object Utils {
+    const val BASE_URL = "https://rickandmortyapi.com/api/"
+}
