+++ movies/src/main/res/layout/activity_movie.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.views.MovieActivity">
+    tools:context=".ui.views.activities.MovieActivity">
 
     <fragment
         android:id="@+id/navHostFragment"
